% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSampleSizes.R
\name{calcSampleSizes}
\alias{calcSampleSizes}
\title{Calculate sample sizes from count or "size" composition data across a set of factors}
\usage{
calcSampleSizes(dfr, id.value = "count", id.facs = "",
  verbose = FALSE)
}
\arguments{
\item{dfr}{- input dataframe with count/abundance data by size and other factors}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{id.facs}{- character vector of column names used as factors in calculating the sample sizes}

\item{verbose}{- flag to print diagnostic output to console}
}
\value{
dataframe with sample sizes across a set of factors.
}
\description{
Function to sample sizes from count or "size" composition data across a set of factors.
}
\details{
Uses \code{sqldf::sqldf}. Note that sample sizes can be calculated from unnormalized size compositions,
as well as "raw" count data, using this function.
}
