% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSizeComps.R, R/getUniqueFZs.R
\name{getUniqueFZs}
\alias{getUniqueFZs}
\title{Determine the unique factor and size combinations from a dataframe of "size" compositions}
\usage{
getUniqueFZs(dfr, id.size = "size", id.value = NULL, id.facs = "",
  cutpts = NULL, expandToAllFactorCombos = FALSE, verbose = verbose)

getUniqueFZs(dfr, id.size = "size", id.value = NULL, id.facs = "",
  cutpts = NULL, expandToAllFactorCombos = FALSE, verbose = verbose)
}
\arguments{
\item{dfr}{- input dataframe with composition data by "size" and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{id.facs}{- character vector of factor column names}

\item{cutpts}{- vector of cutpoints for size compositions (or NULL to extract from dfr)}

\item{expandToAllFactorCombos}{- flag to expand to all factor combinations (TRUE) or only observed combinations (FALSE)}

\item{verbose}{- flag to print diagnostic output to console}

\item{dfr}{- input dataframe with composition data by "size" and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{id.facs}{- character vector of factor column names}

\item{expandToAllFactorCombos}{- flag to expand to all factor combinations (TRUE) or only observed combinations (FALSE)}

\item{verbose}{- flag to print diagnostic output to console}
}
\value{
dataframe with compositions by size (and possibly other factors) expanded to all sizes.

dataframe with compositions by size (and possibly other factors) expanded to all sizes.
}
\description{
Function to determine the unique factor and size combinations from a dataframe of "size" compositions.

Function to determine the unique factor and size combinations from a dataframe of "size" compositions.
}
\details{
If id.facs is blank, all columns other than the size and value columns
are assumed to be factor columns.

Non-standard column names SHOULD NOT be backquoted.

If id.facs is blank, all columns other than the size and value columns
are assumed to be factor columns.

Non-standard column names SHOULD NOT be backquoted.
}
