% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSizeComps.R
\name{normalizeSizeComps}
\alias{normalizeSizeComps}
\title{Normalize "size" compositions across a set of factors}
\usage{
normalizeSizeComps(
  dfr,
  id.size = "size",
  id.value = "count",
  id.facs = "",
  id.normfacs = "",
  verbose = FALSE
)
}
\arguments{
\item{dfr}{- input dataframe with count/abundance data by size and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{id.facs}{- character vector of column names used as factors in calculating the size compositions}

\item{id.normfacs}{- character vector of column names of factors to normalize the size compositions across (a subset of id.facs)}

\item{verbose}{- flag to print diagnostic output to console}
}
\value{
dataframe with "size" compositions normalized to sum to 1 across a set of factors.
}
\description{
Function to create a dataframe of "size" compositions normalized (summing to 1) across a set of factors.
}
\details{
Note that unnormalized size compositions, as well as "raw" count data, can be
converted to normalized size compositions using this function.
}
