% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebinSizeComps.R
\name{rebinSizeComps}
\alias{rebinSizeComps}
\title{Re-bin "size" compositions in a dataframe to all given "size" values (and possibly all factor level combinations)}
\usage{
rebinSizeComps(
  dfr,
  id.size = "size",
  id.value = NULL,
  id.facs = "",
  cutpts = NULL,
  truncate.low = TRUE,
  truncate.high = FALSE,
  expandToAllFactorCombos = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dfr}{- input dataframe with composition data by "size" and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{id.facs}{- character vector of factor column names}

\item{cutpts}{- vector of cutpoints to use for re-binning size compositions}

\item{truncate.low}{-flag to truncate size compositions below the minimum size (TRUE) or include in smallest size bin (FALSE)}

\item{truncate.high}{- flag to truncate size compositions above the maximum size (TRUE) or include in largest size bin (FALSE)}

\item{expandToAllFactorCombos}{- flag to expand to all factor combinations (TRUE) or only observed combinations (FALSE)}

\item{verbose}{- flag to print diagnostic output to console}
}
\value{
dataframe with compositions by size (and possibly other factors) expanded to all sizes by either unique or all factor combinations.
}
\description{
Function to re-bin "size" compositions in a dataframe to all given "size" values (and possibly all factor level combinations).
}
\details{
Non-standard column names SHOULD NOT be back-quoted.
}
