% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleSizeComps.R
\name{scaleSizeComps}
\alias{scaleSizeComps}
\title{Scale "size" compositions in a dataframe by factor level-specific scalars}
\usage{
scaleSizeComps(dfrZCs, dfrScalars, id.value = "fraction",
  id.scalefacs = "", id.scalevalue = "abundance", verbose = FALSE)
}
\arguments{
\item{dfrZCs}{- input dataframe with size composition data by size and other factors}

\item{dfrScalars}{- input dataframe with scalars for a subset of the factors in dfrZCs}

\item{id.value}{- name of column in dfrZCs with (possibly normalized) count/abundance values}

\item{id.scalevalue}{- name of column in dfrScalars with the values used to scale the size compositions in dfrZCs}

\item{verbose}{- flag to print diagnostic output to console}

\item{id.scalevars}{- character vector of column names of factors for the values in dfrScalars to scale the size compositions in dfrZCs by}
}
\value{
dataframe with scaled compositions by size (and possibly other factors).
}
\description{
Function to scale "size" compositions in a dataframe by factor level-specific scalars.
}
\details{
The names of columns in dfrScalars used as factors to scale the
composition data in dfrZCs should be the same in both datasets.
}
