% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleSSsToMax.R
\name{scaleSSsToMax}
\alias{scaleSSsToMax}
\title{Scale sample sizes, with a maximum allowed value}
\usage{
scaleSSsToMax(dfrSS, colSS = "ss", sclSS = 1, maxSS = 200)
}
\arguments{
\item{dfrSS}{- dataframe with sample sizes}

\item{colSS}{- name of column with sample sizes}

\item{sclSS}{- sample size to scale by}

\item{maxSS}{- max output sample size allowed}
}
\value{
copy of input dataframe, but with scaled sampled sizes in column \code{colSS}
}
\description{
Function to scale sample sizes, with a maximum allowed value.
}
\details{
The sample sizes in column \code{colSS} are scaled such that
    dfrSS[[colSS]] = min(maxSS,maxSS/sclSS * dfrSS[[colSS]])
}
