% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleSizes.R
\name{addSampleSizes}
\alias{addSampleSizes}
\title{Add sample sizes to "size" compositions in a dataframe}
\usage{
addSampleSizes(dfr, id.size = "size", id.value = NULL,
  verbose = FALSE)
}
\arguments{
\item{dfr}{- input dataframe with composition data by "size" and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfr with count (or abundance) values}

\item{verbose}{- flag to print diagnostic output to console}
}
\value{
dataframe with original size compositions and sample sizes in column "ss".
}
\description{
Function to add sample sizes to "size" compositions in a dataframe.
}
\details{
Non-standard column names SHOULD NOT be back-quoted.
}
