% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSizeComps.R
\name{calcSizeComps}
\alias{calcSizeComps}
\title{Create "size" compositions from a dataframe with count data}
\usage{
calcSizeComps(dfrZ, id.size = "size", id.value = "abundance",
  id.facs = "", cutpts = NULL, truncate.low = FALSE,
  truncate.high = FALSE, expandToAllFactorCombos = FALSE,
  normalize = FALSE, id.normfacs = "", rescale = FALSE,
  dfrScalars = NULL, id.scalevalue, id.scalefacs = id.normfacs,
  verbose = FALSE)
}
\arguments{
\item{dfrZ}{- input dataframe with count/abundance data by size and other factors}

\item{id.size}{- name of column in dfr with "size" data}

\item{id.value}{- name of column in dfrZ with count (or abundance) values}

\item{id.facs}{- character vector of column names to use as factors in calculating size compositions}

\item{truncate.low}{-flag to truncate counts/abundance below the minimum size or include in first size bin (FALSE)}

\item{truncate.high}{- flag to truncate counts/abundance above the maximum size (TRUE) or include in final size bin (FALSE)}

\item{id.normfacs}{- character vector of column names of factors to normalize the size compositions across when scaling to total abundance}

\item{rescale}{- flag to rescale size compositions by factor combination using scalars in dfrA}

\item{dfrScalars}{- [optional] input dataframe with scalars to apply by a subset of the factors in dfr}

\item{id.scalefacs}{- character vector of column names of factors in dfrA to scale the size compositions across}

\item{verbose}{- flag to print diagnostic output to console}

\item{minSize}{- min "size" cutpoint for binning counts/abundance}

\item{maxSize}{- max "size" cutpoint for binning counts/abundnce}

\item{binSize}{- bin "size"}

\item{id.scalevalaue}{- column name for scale value in dfrA}
}
\value{
dataframe with summed counts (or abundance) as compositions by size (and possibly other factors).
}
\description{
Function to create a dataframe of "size" compositions from a dataframe with count data.
}
\details{
Non-standard column names should be back-quoted. The names of columns used to normalize the composition data in dfrZ
prior to expanding to total abundance using dfrA (id.normfacs) should represent the same factors and levels in both datasets.
}
